book_store/requirements.txt
djangorestframework
django-filter
celery 
redis 
django-celery-beat
gunicorn
python-dotenv
